Score tokenizerできた

+++ src/OtomadUtil/Core/ScoreToken.cs
@@ -5,10 +5,10 @@ using System.Collections.Generic;
 namespace OtomadUtil.Core {
 	public class ScoreToken {
 		public TokenType type;
-		public int actualFrameLength;
-		public int beatLength;
+		public double actualFrameLength;
+		public double beatLength;
 		
-		public ScoreToken(TokenType tt, int af, int bl) {
+		public ScoreToken(TokenType tt, double af, double bl) {
 			type = tt;
 			actualFrameLength = af;
 			beatLength = bl;
